requirements.txt
Django
graphene-django
django-cors-headers
python-dotenv
django-filter
ShopifyAPI
requests
python-slugify
